Move ZD strategies to their own file
